Add `freestyle.tagless` package
Move files in freestyle-tagless modules to freestyle/tagless/ folders.

Remove macro debug println
