+++ freestyle-tagless/js/src/test/scala/freestyle/tagless/taglessTests.scala
@@ -14,9 +14,12 @@
  * limitations under the License.
  */
 
+package freestyle.tagless
+
 import org.scalatest.{Matchers, WordSpec}
 
-import cats.~>
+import scala.scalajs.js.JavaScriptException
+
 import cats.free.Free
 import cats.instances.option._
 
@@ -24,26 +27,19 @@ import algebras._
 import handlers._
 import utils._
 
-class taglessTestsJVM extends WordSpec with Matchers {
+class taglessTestsJS extends WordSpec with Matchers {
 
   "Tagless final algebras" should {
 
+    //Can't catch `StackOverflowError` in scala.js https://github.com/scala-js/scala-js/issues/212
     "blow up the stack when interpreted to stack unsafe monads" in {
-      assertThrows[StackOverflowError] {
+      assertThrows[JavaScriptException] {
         SOProgram[Option](0)
       }
     }
 
     "remain stack safe when interpreted to stack safe monads" in {
-      implicit def liftFree[F[_]]: F ~> Free[F, ?] = λ[F ~> Free[F, ?]](Free.liftF(_))
-
       SOProgram[Free[Option, ?]](0).runTailRec shouldBe Option(iterations)
-
-      type StackSafe[F[_]] = {
-        type λ[α] = Free[F, α]
-      }
-
-      SOProgram[StackSafe[Option]#λ](0).runTailRec shouldBe Option(iterations)
     }
 
   }
